refactor: merge multiComboDetection.ts into multiComboAnalysis.ts
Consolidate multi-combo related functions into a single module:

- Move detectLeadingMultiCombo function from multiComboDetection.ts to multiComboAnalysis.ts
- Update all imports to reference the consolidated module
- Remove the now-empty multiComboDetection.ts file
- Add MultiComboDetectionResult interface to multiComboAnalysis.ts

This reduces the number of modules and keeps all multi-combo functionality together.
All tests continue to pass with the expected multiComboTrumpStructureMatching failure.
